requirements.txt
Click==7.0
Flask==1.1.1
py2neo==4.3.0
google-cloud-datastore==1.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
